attack and defend

+++ RolePlayingGame/Interfaces/IWearableItem.cs
@@ -5,8 +5,8 @@ namespace RolePlayingGame.Interfaces
     public interface IWearableItem
     {
          WearableItemType ItemType { get;}
-         double AttackPoints { get; set; }
-         double DefencePoints { get; set; }
+         int AttackPoints { get; set; }
+         int DefencePoints { get; set; }
          double CritChance { get; set; }
          double CritDmg { get; set; }
          double DodgeChance { get; set; }
